created About folder from page (#1018)
* Separate About component into index.js and styles.js

* Fix merge conflicts
