+++ src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom';
 import { QueryParamProvider } from 'use-query-params';
 import Layout from './components/common/Layout';
 import ScrollToTop from './components/common/ScrollToTop';
-import About from './pages/About';
+import About from './pages/About/index';
 import Contact from './pages/Contact';
 import Collaborate from './pages/Collaborate';
 import Contributors from './pages/Contributors';
